Fix Windows build
probably break linux (fine for now)

+++ source/timer/Timer.cpp
@@ -13,7 +13,7 @@ void Timer::Start()
 std::string Timer::Result()
 {
 	const auto end = std::chrono::high_resolution_clock::now();
-	const int ticks = std::chrono::duration<double, std::milli>(end - m_start).count();
+	const double ticks = std::chrono::duration<double, std::milli>(end - m_start).count();
 	const std::string time = std::to_string(ticks) + "ms";
 	return time;
 }
